Made code for the connection with the arduino with the rfid reader, and code that sends the speed to the server. Divided code in smaller parts, so everything is easier to read.

+++ rp6control/Setters.ino
@@ -25,7 +25,7 @@ void setTurningAngle(int angle) {
 }
 
 bool checkAndSetController(const String sCommand, int clientId) {
-  int commands[2];
+  long commands[2];
   if (!parseCommand(sCommand, commands) != 0) {
     return false;
   }
